@ feat: OGG Vorbis support, exclusive fullscreen, official offset timing
- Audio: enable OGG Vorbis via stb_vorbis (include before miniaudio.h)
- Display: exclusive fullscreen checkbox + --exclusive/--no-exclusive CLI
  + Alt+Enter runtime toggle
- Offset: official formula - music from 0, planet pre-rotates offset_sec
  degrees at initial BPM, reaches tile 1 (not tile 0)
- preRoll: minimum 360° rotation on tile 0, audio delayed to match
- Fix: group splitting (1M max) prevents malloc fail on 6.7M tile levels
- Fix: buildGrid world-extent cap (1024x1024) prevents grid explosion
- Fix: selected-tile playback starts at exact tile (missing countdown)
- Fix: double-precision trail pipeline (Catmull-Rom in float64)
- Perf: skip decorations JSON parsing entirely
- Perf: window visible immediately (swap before mesh build)

@

+++ game/Planet.cpp
@@ -44,7 +44,7 @@ void Planet::update(float currentTime) {
     (void)currentTime;
 }
 
-void Planet::setTrailPoints(const float* xy, int count) {
+void Planet::setTrailPoints(const double* xy, int count) {
     if (trail)
         trail->setPoints(xy, count);
 }
